requirements.txt
coreapi==1.32.0
Django==1.10
djangorestframework==3.4.4
gunicorn==19.6.0
itypes==1.1.0
psycopg2==2.6.2
requests==2.11.0
simplejson==3.8.2
uritemplate==0.6
